Refactor organize subsearch GUI package and update module names
This commit refactors the GUI portion of the subsearch package by reorganizing some of its modules and updating their names. The following changes were made:

    Moved tab_application_settings.py and tab_search_settings.py files from src/subsearch/gui to src/subsearch/gui/tabs directory.
    Renamed tab_manual_download.py to dowload_tab.py and tab_languages.py to language_tab.py in src/subsearch/gui/tabs directory.
    Deleted tk_tools.py and widget_menu.py files from src/subsearch/gui directory.
    Created new tabs subdirectory in src/subsearch/gui and added new __init__.py file in src/subsearch/gui/tabs directory.
    Modified search_tab.py, settings_tab.py, tab_manager.py, tkinter_utils.py, core.py, and __init__.py files in src/subsearch and src/subsearch/gui/tabs directories to reflect the changes in module names and directory structure.

These changes are intended to improve the organization and readability of the GUI portion of the subsearch package.
